requirements.txt
# Basic requirements for E-ZPass Fraud Detection
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-auth>=2.17.0
pandas>=1.5.0
python-dotenv>=1.0.0